+++ PillBoxWebAPI/pillboxApp/Tables/MedicationSchedule.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[MedicationSchedule]
 (
-	[Id] INT NOT NULL IDENTITY(1,1) PRIMARY KEY, 
-    [MedicationId] INT NOT NULL, 
+	[Id] BIGINT NOT NULL IDENTITY(1,1) PRIMARY KEY, 
+    [MedicationId] BIGINT NOT NULL, 
     [StartDate] DATETIME NOT NULL, 
     [RepeatNumber] INT NOT NULL, 
 	[TimeFrame] NVARCHAR(255) NOT NULL,
